+++ example_test.go
@@ -20,9 +20,9 @@ func Example() {
 	if err != nil {
 		log.Panic(err)
 	}
-	fmt.Printf("Portuguese (BR) city name: %v\n", record.City.Names["pt-BR"])
-	fmt.Printf("English subdivision name: %v\n", record.Subdivisions[0].Names["en"])
-	fmt.Printf("Russian country name: %v\n", record.Country.Names["ru"])
+	fmt.Printf("Portuguese (BR) city name: %v\n", record.City.Names.BrazilianPortuguese)
+	fmt.Printf("English subdivision name: %v\n", record.Subdivisions[0].Names.English)
+	fmt.Printf("Russian country name: %v\n", record.Country.Names.Russian)
 	fmt.Printf("ISO country code: %v\n", record.Country.ISOCode)
 	fmt.Printf("Time zone: %v\n", record.Location.TimeZone)
 	fmt.Printf("Coordinates: %v, %v\n", record.Location.Latitude, record.Location.Longitude)